Move RequestToDtoSourceEnum to Attribute\Enum namespace
Follow CakePHP enum convention from #19348:
- Enums in Enum/ subdirectory
- No Enum suffix on class name

Cake\Controller\Attribute\RequestToDtoSourceEnum
-> Cake\Controller\Attribute\Enum\RequestToDtoSource
